Partially revert "store: allow shifting only with contiguous mappings"
We'd like to use ID-shifting of layers in order to prevent a copy/chown
operation when --userns keep-id is used, and are able to work around the
problems this causes elsewhere.

This is a partial revert of commit cd673d561192e35ac657dddd13ed57ef0b1bfff6
in github.com/containers/storage.